Remove old FilterIndexRule & JoinIndexRule and getCandidateIndexes
