+++ src/main/scala/com/microsoft/hyperspace/index/rules/RuleUtils.scala
@@ -38,130 +38,6 @@ import com.microsoft.hyperspace.util.HyperspaceConf
 object RuleUtils {
 
   /**
-   * Filter the given candidate indexes by matching signatures and index status.
-   * If Hybrid Scan is enabled, it compares the file metadata directly, and does not
-   * match signatures. By doing that, we could perform file-level comparison between
-   * index source files and the input files of the given plan. If there are some common
-   * files, the index is considered as a candidate.
-   *
-   * @param spark Spark Session.
-   * @param indexes List of available indexes.
-   * @param relation Relation with which given indexes are compared.
-   * @return Active indexes built for this plan.
-   */
-  def getCandidateIndexes(
-      spark: SparkSession,
-      indexes: Seq[IndexLogEntry],
-      relation: FileBasedRelation): Seq[IndexLogEntry] = {
-    // Map of a signature provider to a signature generated for the given plan.
-    val signatureMap = mutable.Map[String, Option[String]]()
-
-    def signatureValid(entry: IndexLogEntry): Boolean = {
-      entry.withCachedTag(relation.plan, IndexLogEntryTags.SIGNATURE_MATCHED) {
-        val sourcePlanSignature = entry.signature
-
-        signatureMap.getOrElseUpdate(
-          sourcePlanSignature.provider,
-          LogicalPlanSignatureProvider
-            .create(sourcePlanSignature.provider)
-            .signature(relation.plan)) match {
-          case Some(s) => s.equals(sourcePlanSignature.value)
-          case None => false
-        }
-      }
-    }
-
-    val hybridScanEnabled = HyperspaceConf.hybridScanEnabled(spark)
-    val hybridScanDeleteEnabled = HyperspaceConf.hybridScanDeleteEnabled(spark)
-
-    def getHybridScanCandidate(index: IndexLogEntry): Option[IndexLogEntry] = {
-      // TODO: As in [[PlanSignatureProvider]], Source plan signature comparison is required to
-      //  support arbitrary source plans at index creation.
-      //  See https://github.com/microsoft/hyperspace/issues/158
-
-      val entry = relation.closestIndex(index)
-
-      val isHybridScanCandidate =
-        entry.withCachedTag(relation.plan, IndexLogEntryTags.IS_HYBRIDSCAN_CANDIDATE) {
-          // Find the number of common files between the source relation and index source files.
-          // The total size of common files are collected and tagged for candidate.
-          val (commonCnt, commonBytes) = relation.allFileInfos.foldLeft(0L, 0L) { (res, f) =>
-            if (entry.sourceFileInfoSet.contains(f)) {
-              (res._1 + 1, res._2 + f.size) // count, total bytes
-            } else {
-              res
-            }
-          }
-
-          val appendedBytesRatio = 1 - commonBytes / relation.allFileSizeInBytes.toFloat
-          val deletedBytesRatio = 1 - commonBytes / entry.sourceFilesSizeInBytes.toFloat
-
-          val deletedCnt = entry.sourceFileInfoSet.size - commonCnt
-          val isAppendAndDeleteCandidate = hybridScanDeleteEnabled && entry.hasLineageColumn &&
-            commonCnt > 0 &&
-            appendedBytesRatio < HyperspaceConf.hybridScanAppendedRatioThreshold(spark) &&
-            deletedBytesRatio < HyperspaceConf.hybridScanDeletedRatioThreshold(spark)
-
-          // For append-only Hybrid Scan, deleted files are not allowed.
-          lazy val isAppendOnlyCandidate = deletedCnt == 0 && commonCnt > 0 &&
-            appendedBytesRatio < HyperspaceConf.hybridScanAppendedRatioThreshold(spark)
-
-          val isCandidate = isAppendAndDeleteCandidate || isAppendOnlyCandidate
-          if (isCandidate) {
-            entry.setTagValue(
-              relation.plan,
-              IndexLogEntryTags.COMMON_SOURCE_SIZE_IN_BYTES,
-              commonBytes)
-
-            // If there is no change in source dataset, the index will be applied by
-            // transformPlanToUseIndexOnlyScan.
-            entry.setTagValue(
-              relation.plan,
-              IndexLogEntryTags.HYBRIDSCAN_REQUIRED,
-              !(commonCnt == entry.sourceFileInfoSet.size
-                && commonCnt == relation.allFileInfos.size))
-          }
-          isCandidate
-        }
-      if (isHybridScanCandidate) {
-        Some(entry)
-      } else {
-        None
-      }
-    }
-
-    def prepareHybridScanCandidateSelection(
-        spark: SparkSession,
-        plan: LogicalPlan,
-        indexes: Seq[IndexLogEntry]): Unit = {
-      assert(HyperspaceConf.hybridScanEnabled(spark))
-      val curConfigs = Seq(
-        HyperspaceConf.hybridScanAppendedRatioThreshold(spark).toString,
-        HyperspaceConf.hybridScanDeletedRatioThreshold(spark).toString)
-
-      indexes.foreach { index =>
-        val taggedConfigs = index.getTagValue(plan, HYBRIDSCAN_RELATED_CONFIGS)
-        if (taggedConfigs.isEmpty || !taggedConfigs.get.equals(curConfigs)) {
-          // Need to reset cached tags as these config changes can change the result.
-          index.unsetTagValue(plan, IS_HYBRIDSCAN_CANDIDATE)
-          index.setTagValue(plan, HYBRIDSCAN_RELATED_CONFIGS, curConfigs)
-        }
-      }
-    }
-
-    if (hybridScanEnabled) {
-      // TODO: Duplicate listing files for the given relation as in
-      //  [[transformPlanToUseHybridScan]]
-      //  See https://github.com/microsoft/hyperspace/issues/160
-      prepareHybridScanCandidateSelection(spark, relation.plan, indexes)
-
-      indexes.filter(_.created).flatMap(getHybridScanCandidate)
-    } else {
-      indexes.filter(index => index.created && signatureValid(index))
-    }
-  }
-
-  /**
    * Check if an index was applied the given relation or not.
    * This can be determined by an identifier in [[FileBasedRelation]]'s options.
    *
