+++ src/test/scala/com/microsoft/hyperspace/index/rules/RuleUtilsTest.scala
@@ -87,22 +87,6 @@ class RuleUtilsTest extends HyperspaceRuleSuite with SQLHelper {
     createIndexLogEntry("t2i2", Seq(t2c1, t2c2), Seq(t2c3), t2ProjectNode)
   }
 
-  test("Verify indexes are matched by signature correctly.") {
-    val indexManager = IndexCollectionManager(spark)
-    val allIndexes = indexManager.getIndexes(Seq(Constants.States.ACTIVE))
-
-    val t1Relation = RuleUtils.getRelation(spark, t1ProjectNode).get
-    val t2Relation = RuleUtils.getRelation(spark, t2ProjectNode).get
-    assert(RuleUtils.getCandidateIndexes(spark, allIndexes, t1Relation).length === 3)
-    assert(RuleUtils.getCandidateIndexes(spark, allIndexes, t2Relation).length === 2)
-
-    // Delete an index for t1ProjectNode
-    indexManager.delete("t1i1")
-    val allIndexes2 = indexManager.getIndexes(Seq(Constants.States.ACTIVE))
-
-    assert(RuleUtils.getCandidateIndexes(spark, allIndexes2, t1Relation).length === 2)
-  }
-
   test("Verify get logical relation for single logical relation node plan.") {
     validateLogicalRelation(t1ScanNode, t1ScanNode)
   }
@@ -115,202 +99,6 @@ class RuleUtilsTest extends HyperspaceRuleSuite with SQLHelper {
     val joinNode = JoinWithoutHint(t1ProjectNode, t2ProjectNode, JoinType("inner"), None)
     val r = RuleUtils.getRelation(spark, Project(Seq(t1c3, t2c3), joinNode))
     assert(r.isEmpty)
-  }
-
-  test("Verify getCandidateIndex for hybrid scan") {
-    withTempPath { tempPath =>
-      val indexManager = IndexCollectionManager(spark)
-      val df = spark.range(1, 5).toDF("id")
-      val dataPath = tempPath.getAbsolutePath
-      df.write.parquet(dataPath)
-
-      val indexNameWithLineage = "index1"
-      val indexNameWithoutLineage = "index2"
-      withIndex(indexNameWithLineage, indexNameWithoutLineage) {
-        val readDf = spark.read.parquet(dataPath)
-        val expectedCommonSourceBytes = FileUtils.getDirectorySize(new Path(dataPath))
-        withSQLConf(IndexConstants.INDEX_LINEAGE_ENABLED -> "true") {
-          indexManager.create(readDf, IndexConfig(indexNameWithLineage, Seq("id")))
-        }
-        withSQLConf(IndexConstants.INDEX_LINEAGE_ENABLED -> "false") {
-          indexManager.create(readDf, IndexConfig(indexNameWithoutLineage, Seq("id")))
-        }
-        val allIndexes = indexManager.getIndexes(Seq(Constants.States.ACTIVE))
-
-        def verify(
-            plan: LogicalPlan,
-            hybridScanEnabled: Boolean,
-            hybridScanDeleteEnabled: Boolean,
-            expectedCandidateIndexes: Seq[String],
-            expectedHybridScanTag: Option[Boolean],
-            expectedCommonSourceBytes: Option[Long]): Unit = {
-          withSQLConf(
-            IndexConstants.INDEX_HYBRID_SCAN_ENABLED -> hybridScanEnabled.toString,
-            IndexConstants.INDEX_HYBRID_SCAN_APPENDED_RATIO_THRESHOLD -> "0.99",
-            IndexConstants.INDEX_HYBRID_SCAN_DELETED_RATIO_THRESHOLD ->
-              (if (hybridScanDeleteEnabled) "0.99" else "0")) {
-            val relation = RuleUtils.getRelation(spark, plan).get
-            val indexes = RuleUtils
-              .getCandidateIndexes(spark, allIndexes, relation)
-            if (expectedCandidateIndexes.nonEmpty) {
-              assert(indexes.length === expectedCandidateIndexes.length)
-              assert(indexes.map(_.name).toSet.equals(expectedCandidateIndexes.toSet))
-              indexes.foreach { index =>
-                assert(
-                  index.getTagValue(plan, IndexLogEntryTags.HYBRIDSCAN_REQUIRED)
-                    === expectedHybridScanTag)
-                assert(
-                  index.getTagValue(plan, IndexLogEntryTags.COMMON_SOURCE_SIZE_IN_BYTES)
-                    === expectedCommonSourceBytes)
-              }
-            } else {
-              assert(indexes.isEmpty)
-            }
-          }
-        }
-
-        // Verify that a candidate index is returned with the unmodified data files whether
-        // hybrid scan is enabled or not.
-        {
-          val optimizedPlan = spark.read.parquet(dataPath).queryExecution.optimizedPlan
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = false,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(indexNameWithLineage, indexNameWithoutLineage),
-            expectedHybridScanTag = None,
-            expectedCommonSourceBytes = None)
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = true,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(indexNameWithLineage, indexNameWithoutLineage),
-            expectedHybridScanTag = Some(false),
-            expectedCommonSourceBytes = Some(expectedCommonSourceBytes))
-        }
-
-        // Scenario #1: Append new files.
-        df.write.mode("append").parquet(dataPath)
-
-        {
-          val optimizedPlan = spark.read.parquet(dataPath).queryExecution.optimizedPlan
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = false,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(),
-            expectedHybridScanTag = None,
-            expectedCommonSourceBytes = None)
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = true,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(indexNameWithLineage, indexNameWithoutLineage),
-            expectedHybridScanTag = Some(true),
-            expectedCommonSourceBytes = Some(expectedCommonSourceBytes))
-        }
-
-        // Scenario #2: Delete 1 file.
-        val deleteFilePath = new Path(readDf.inputFiles.head)
-        val updatedExpectedCommonSourceBytes = expectedCommonSourceBytes - FileUtils
-          .getDirectorySize(deleteFilePath)
-        FileUtils.delete(deleteFilePath)
-
-        {
-          val optimizedPlan = spark.read.parquet(dataPath).queryExecution.optimizedPlan
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = false,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(),
-            expectedHybridScanTag = None,
-            expectedCommonSourceBytes = None)
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = true,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(),
-            expectedHybridScanTag = None,
-            expectedCommonSourceBytes = None)
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = true,
-            hybridScanDeleteEnabled = true,
-            expectedCandidateIndexes = Seq(indexNameWithLineage),
-            expectedHybridScanTag = Some(true),
-            expectedCommonSourceBytes = Some(updatedExpectedCommonSourceBytes))
-        }
-
-        // Scenario #3: Replace all files.
-        df.write.mode("overwrite").parquet(dataPath)
-
-        {
-          val optimizedPlan = spark.read.parquet(dataPath).queryExecution.optimizedPlan
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = false,
-            hybridScanDeleteEnabled = false,
-            expectedCandidateIndexes = Seq(),
-            expectedHybridScanTag = None,
-            expectedCommonSourceBytes = None)
-          verify(
-            optimizedPlan,
-            hybridScanEnabled = true,
-            hybridScanDeleteEnabled = true,
-            expectedCandidateIndexes = Seq(),
-            expectedHybridScanTag = None,
-            expectedCommonSourceBytes = None)
-        }
-      }
-    }
-  }
-
-  test("Verify Hybrid Scan candidate tags work as expected.") {
-    withTempPath { tempPath =>
-      val indexManager = IndexCollectionManager(spark)
-      val df = spark.range(1, 10).toDF("id")
-      val dataPath = tempPath.getAbsolutePath
-      df.write.parquet(dataPath)
-
-      withIndex("index1") {
-        val readDf = spark.read.parquet(dataPath)
-        withSQLConf(IndexConstants.INDEX_LINEAGE_ENABLED -> "true") {
-          indexManager.create(readDf, IndexConfig("index1", Seq("id")))
-        }
-
-        val allIndexes = indexManager.getIndexes(Seq(Constants.States.ACTIVE))
-        df.limit(5).write.mode("append").parquet(dataPath)
-        val optimizedPlan = spark.read.parquet(dataPath).queryExecution.optimizedPlan
-        val relation = RuleUtils.getRelation(spark, optimizedPlan).get
-
-        withSQLConf(IndexConstants.INDEX_HYBRID_SCAN_ENABLED -> "true") {
-          withSQLConf(IndexConstants.INDEX_HYBRID_SCAN_APPENDED_RATIO_THRESHOLD -> "0.99") {
-            val indexes = RuleUtils.getCandidateIndexes(spark, allIndexes, relation)
-            assert(
-              indexes.head
-                .getTagValue(relation.plan, IndexLogEntryTags.IS_HYBRIDSCAN_CANDIDATE)
-                .get)
-            assert(
-              indexes.head
-                .getTagValue(relation.plan, IndexLogEntryTags.HYBRIDSCAN_RELATED_CONFIGS)
-                .get == Seq("0.99", "0.2"))
-          }
-
-          withSQLConf(IndexConstants.INDEX_HYBRID_SCAN_APPENDED_RATIO_THRESHOLD -> "0.2") {
-            val indexes = RuleUtils.getCandidateIndexes(spark, allIndexes, relation)
-            assert(indexes.isEmpty)
-            assert(
-              !allIndexes.head
-                .getTagValue(relation.plan, IndexLogEntryTags.IS_HYBRIDSCAN_CANDIDATE)
-                .get)
-            assert(
-              allIndexes.head
-                .getTagValue(relation.plan, IndexLogEntryTags.HYBRIDSCAN_RELATED_CONFIGS)
-                .get == Seq("0.2", "0.2"))
-          }
-        }
-      }
-    }
   }
 
   test("Verify the location of injected shuffle for Hybrid Scan.") {
